refactor: Move validators to file
Move custom class validators to a dedicated file 'validators.py' to increase code modularity.
Rename static functions for semantics and readability.
Ensure tests pass after refactor.
