+++ greetings/views.py
@@ -1,5 +1,3 @@
-import re
-
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
@@ -7,7 +5,7 @@ from rest_framework.request import Request
 
 from greetings.models import Greeting
 from greetings.serializers import GreetingSerializer
-
+from greetings.utils.validators import * 
 
 @api_view(["GET"])
 def list_greetings(request: Request) -> Response:
@@ -26,7 +24,7 @@ def save_custom_greeting(request: Request) -> Response:
   """Save a custom greeting from a user."""
 
   custom_greeting = validate_query_param(request)
-  
+
   try:    
     greeting = Greeting(greeting_text=custom_greeting)
     greeting.save()
@@ -39,30 +37,9 @@ def save_custom_greeting(request: Request) -> Response:
 
 
 def validate_query_param(request: Request) -> str:
-  GreetingPathValidator.validate_query_param_present(request)
+  GreetingPathValidator.validate_param_key_present(request)
   custom_greeting = request.query_params['greeting']
   GreetingParamValidator.validate_not_blank(custom_greeting)
-  GreetingValueValidator.validate_query_param_value(custom_greeting)
+  GreetingValueValidator.validate_param_value(custom_greeting)
   return custom_greeting
-
-class GreetingPathValidator:
-  @staticmethod
-  def validate_query_param_present(request: Request) -> None:
-    query_param_key = '?greeting='
-    url_path = request.build_absolute_uri()
-    if query_param_key not in url_path:
-      raise ValueError("Key for required query param: greeting is missing.")
-
-class GreetingParamValidator:  
-  @staticmethod
-  def validate_not_blank(value: str) -> None:
-    if value.strip() == '':
-      raise ValueError("Value for query param: greeting cannot be blank or empty.")
-
-class GreetingValueValidator:
-  @staticmethod
-  def validate_query_param_value(value: str) -> None:
-    pattern = r'^[a-zA-Z]+$'
-    if not re.match(pattern, value):
-      raise ValueError("Value for query param: greeting cannot contain non alphabetic chars.")
